Add feature flag for application tracker
