+++ src/feature-flags.js
@@ -1,0 +1,12 @@
+const featureFlags = [
+  {
+    name: 'userProfile',
+    isActive: false,
+  },
+  {
+    name: 'applicationTracker',
+    isActive: true,
+  },
+];
+
+export default featureFlags;
